shell/panelview: Use the panel geometry as a fallback for the mask
It seems like in some scenarios (e.g. detaching and reattaching monitors) ksvg provides
an empty mask until the panel is re-drawn; as a first workaround, we add a fallback that
avoids the panel spilling into other monitors when an empty mask is provided.

BUG: 481380
FIXED-IN: 6.1


(cherry picked from commit 2a125115418869eeb62c0dab8b7069202bf1fd5b)

b52a2dc8 Use the panel geometry as a fallback for the mask

Co-authored-by: Niccolò Venerandi <niccolo@venerandi.com>