applets/lock_logout: add "show logout screen" option
With the change to simplify the logout greeter UI when invoked with a specific option, there is currently no option to trigger the logout greeter with all of the options via any existing actions.

This adds a logout screen action which shows the logout greeter with available options.


FEATURE: 488832
FIXED-IN: 6.2