containmentactions/contextmenu: rephrase ambiguous "Leave…" action
This action now does a very specific thing: it shows the logout screen
with all options — in contrast to all other actions which show it in
confirmation mode, with only the selected option and a cancel button.

Let's rephrase the menu item to reflect that, since we can now use
precise language.

BUG: 489179
FIXED-IN: 6.2