applets/lock_logout: Restore pre-6.0 UX
### Reason for the change
This widget lets you configure what actions are shown, and currently it
defaults to showing the "Lock" and "Shut Down" actions. In the past 
these made sense because Shut Down showed the logout screen with all
options, and Lock was the only thing not on the logout screen. So you 
had the option to access everything.

But in Plasma 6, Shut Down won't show the full logout screen with all
options; it now shows only a confirmation-style screen with just Shut 
Down and Cancel

To restore the old intended UX, let's replace Shut Down with "Show
Logout Screen", from which all options can be chosen.


### Test plan

Add a new instance of the widget -> it gets the "Show Logout Screen" action by default instead of "Shut Down"

Un-apply the MR, add a new instance of the widget, customize it, apply the MR, and restart Plasma -> prior customizations are saved

@teams/usability @tcarmichael