BatteryIcon: let it optionally show a power profile badge
This commit gives BatteryIcon the ability to be given the name of a
power profile icon and append it to the end of the icon name it asks the
icon for. This facilitates work to badge the battery icon with the power
profile when it's in a non-default power profile state.

There is no effect when using an icon theme without the requisite icons;
it'll simply fall back to the status quo.

CCBUG: 483805