applets/cameraindicator: fix building test with pipewire < 0.3.80 again
spa/params/tag-utils.h is not actually used, and breaks the build for
older versions of pipewire, including the one shipped in neon jammy
(current version)

Commit be83f32a fixed this issue for videosource.cpp, but it still
applies to videoplay.cpp. Removing it here as well allows building
plasma-workspace on neon. videoplay.cpp does not seem to actually use
this file.