xembedsniproxy: Warp pointer to click location on wayland
Some clients track pointer focus and ignore clicks if they think
they dont have a window under the cursor. On Wayland pointer focus
is driven by the compositor - the client receives a click without
a prior enter event. To workaround this warp the pointer to the
click position before clicking which generates the needed enter.
BUG:489286
FIXED-IN:6.1.3