Modernize ActionButton code
ActionButton is now an AbstractButton subclass.
This allows it to inherit API, interaction and accessibility related behaviors that most buttons should have.

LogoutButton was also changed since it's an ActionButton subclass.
labelRendering was removed because the linked bug can no longer be reproduced.
circleVisiblity and circleOpacity were removed because they don't seem to make the difference the comment says they make.