applets/sysmon: Base switch size on both compact and full rep size
This reworks the switching size logic so we account for both maximum
compact rep size and minimum full rep size. The intent is that we switch
the moment the compact rep reaches its maximum, unless the full rep
won't fit, in which case we'll switch when the full rep fits. We also
account for if we're on a horizontal or vertical panel, only accounting
for one dimension for switching in those cases, as the other dimension
is free to expand.

BUG: 490108