Modernize UserDelegate
It is now an ItemDelegate subclass with automatic accessibility info

We use implicit size now, so no need to set size externally in order to see everything.

We use Kirigami icon sizes and spacings now.

For naming clarity: name -> displayName, iconSource -> iconName

Opacity animation duration was changed to veryLongDuration to sync
with other animations that usually play at the same time.

All the icon/avatar things that were previously done with an Image
and Kirigami Icon could actually be done with just the Kirigami Icon.
We use one Kirigami Icon for avatars and icons now.

font size is now automatically scaled with the icon size.
The original intention for the font size was to make the
font bigger to go with the icon size visually.