Simplify WallpaperFader
We don't need to plug in a set group of items, 
we can just expose the UI opacity that should be applied and bind to that.

For some reason, the opacity animation duration wasn't the same as the factor animation duration.
I made them the same so that they'd move in sync.