Modernize Clock
Previously, you had to add a `shadow` property to this component
holding the clock's shadow in order for the wallpaper fader to work right.
The clock and clock shadow both had to be managed by the wallpaper fader.
Now you can just bind the wallpaper fader's `uiOpacity` to this component's `uiOpacity`
and you get the same effect with less deeply nested behaviors.

Also, we use MultiEffect instead of the Qt5 compatibility DropShadow.
The appearance is not exactly the same, but it looks fine.