startplasma: Also set environment variable when it is not currently set
qgetenv returns a null QByteArray when there is currently no such variable.
A null QByteArray is distinct from an empty one. However, operator!= will
treat them the same. Therefore it's impossible to set an environment variable
to an empty value.

Explicitly check for null and also qputenv it when this is the case.

BUG: 490432