dataengines/weather/wetter.com: Use location codes for internal mapping
The engine used to store info about the different locations in
a hashmap using the full place name as the key.

Use instead the location code as the key for the different maps,
so this is unique, and different places with the same name and
country doesn't get merged into one.

The format of the weather "source" string is not changed
("wettercom|weather|<place>|<code>;<displayName>")
so that current user's configs do not break.

CCBUG: 490362