notifications: When doing a drag and drop, set the supported action to cpoy
By default QDrag::exec performs a drag where the allowed actions are
only "move".

The mimedata for a screenshot is a URI list to the file, dragging and
dropping should not move the original file.

Equally importantly if we want to share a file, that's effectively a
copy. Some clients will rightfully not accept a URI list with the only
action being move.

Testing done:

Screenshot taken with the spectacle global shortcut, then drag from that
into the Electron app "Discord"

BUG: 469644