kcms/notifications: Set the config paths in the right order
`QStandardPaths` finds the paths following the order of precedence
given by `$XDG_DATA_DIRS, that is, more priority goes first.

It is the opposite for `KConfig::addConfigSources()`, which goes from
more global first to more specific (the latter override the previous
ones), so we need to reverse the order.

This is a niche issue mostly just for dev sessions, but let's
better have it correctly.