kcms/autostart: remove ellipses from menu button
kcm_autostart uses a Kirigami.Action in its header that turns into a
menu button due to its child actions. The Action has the label "Add…".
While the HIG recommend the use of ellipses for buttons with Action
labels that require further user input, this was recently updated to
exclude menu buttons as this is redundant with the downward-pointing
arrow on this control.

This change removes the ellipsis, and adapts the appium test to this.