kcms/autostart: prepare for setting action a11y
The button menu in the header of this kcm uses context to shorten the
labels, which makes them harder to with screen readers. Currently
Kirigami.Actions cannot have an Accessible.name set directly, but Qt 6.8
will support this.

This change adds comments that set the Accessible.name, along with
instructions to uncomment them once Qt 6.8 is required.