applets/mediacontroller: Fix label width overflow in horizontal panel
We need to limit the width of the internal item (`playerRow`) to the
available width of the applet so that the labels don't overflow.

The requested `Layout.preferredWidth` of the applet is still the
`implicitWidth` so the applet's general size calculations shouldn't
have any changes.


(cherry picked from commit d88aff1897c7cc3c6f6e06c40b7a9bc12361f639)

Co-authored-by: Ismael Asensio <isma.af@gmail.com>