klipper: highlight the first item in popup menu
In the old Meta+V QtWidgets menu popup, the top item was always highlighted, so
you could search until the item you wanted was on top, and then immediately hit
the return key.

With the new QML-based one, you have to hit the down arrow before the return
key. It's a tiny bit slower and more annoying.

This restores the old behavior.

BUG: 491142
FIXED-IN: 6.2