kcms/region_language: run "localectl list-locales" without pager
<!-- Thank you for submitting a merge request to make KDE Software better! Use the information from the commit message(s) to fill in the template below.

### Original commit message(s)

* kcms/region_language: run "localectl list-locales" without pager

(For future reference, follow the guidelines at https://cbea.ms/git-commit/#seven-rules)
-->


### Reason for the change

<!--
If this merge request change is resolving a problem, describe the problem and explain why the approach taken here is the correct way to resolve it.

If the merge request adds a new feature or UI change, explain the benefits of the change.
-->
By default, `localectl` shows its output via a pager. This can be configured (incl. disabled) with the environment variable `SYSTEMD_PAGER`. The typically used pagers (`less` or `more`) paginate only if run in terminal. However, this cannot be guaranteed. If the output is paginated, the code could not parse the `localectl` output as expected.

### Test plan

<!--
Extend any related autotests with new test cases to validate your changes. If none exist, consider adding one if you have the technical skills to do so.

Otherwise, at least describe how you tested the change, and how someone reviewing this merge request can test it themselves. No need to mention obvious things like "Apply the changes".
-->
I don't know how to test this.