applets/notifications: Base popup width on content width
This removes the fixed width constraint of notification popups, instead
basing the width on its content width. It uses the old popup width as
minimum width and allows the popup to expand to a maximum width that is
currently defined as 3 * the minimum width, unless the screen is smaller
in which case it will be limited to screen size. This allows the popup
to better accomodate large or many actions.

BUG: 490582

min max width