applets/notifications: Fix placement of NotificationItem in NotificationPopup
Rather than relying on the item being placed at 0,0 with full width and
then correcting for fact that certain elements overlay elements from the
popup in NotificationItem, use anchors for horizontal placement of
NotificationItem and account for underlying items in the anchor margins.
While we're at it, also make sure this works correctly in RtL mode.