Don't propagate snap confinement related variables to other environments
This leads to bad side-effects. A process started via systemd StartUnit
would carry the confinement related variables of another process. Due to
this, Qt would for instance load the wrong platform theme. For instance,
this can lead to bad deadlocks like kwin requiring the desktop portal,
while the desktop portal waits on kwin to startup.

So better filter out variables known to be related to snap confinements.
They are supposed to be process specific anyway.