lookandfeel: Support the "osdAdditionalText" property in fallback OsdItem
This is currently only used by the mediaPlayerVolumeChanged method,
where `playerName` is relayed to the OSD as additional text.
We can make use of this also for an upcoming screenBrightnessChanged
method addition.

In the QML, the label item is pulled upwards in the code so it can
appear on top of the progress bar, instead of being mutually
exclusive. Preferred height is adjusted to also take implicitHeight
into account, rather than only the size of the icon.