shell: First stab at an OSD screenBrightnessChanged method
The old brightnessChanged method with its lone percentage argument
isn't sufficient in a per-screen brightness world. The new addition
adds several arguments to identify and present brightness for
different displays.

Each method call receives information about a single screen,
which will be combined with information from subsequent calls.
Brightness info about all known displays will be retained until
the OSD disappears.

As a first stab, we use a multi-line text OSD indicator when more
than one screen sees a brightness change. This is specialized to use
`showProgress()` when it's only a single screen, using a progress
indicator with the newly unbroken `osdAdditionalText` property for
the QML `Osd` item to show which screen is involved.

As a further specialization, the original progress indicator without
additional text is used when only one screen changed brightness and
could be associated to the only screen used by the ShellCorona.

With all this in place, we can keep basic single-screen changes
simple and straightforward, while being able to present relevant
changes for a larger number of screens. Eventually, we should
probably create multiple OSD popups on associated screens whenever
it's possible to determine the correct QScreen for each call and
place OSD dialogs accordingly.