Add Install Updates & Shut Down feature to Logout page
Depends on https://invent.kde.org/frameworks/breeze-icons/-/merge_requests/392

This feature reboots the computer, does an offline update, then shuts
down after the offline update has completed, rather than booting into
the operating system again.

The idea with this is that you can set your computer to update and
shut down and then walk away. Currently, there's no way to do this in
Plasma, so you can't leave the computer to update when you're done
with it; you have to wait for it to update then turn it off again.

Windows has similar functionality that this emulates, but this does
have to reboot the computer into the offline updater, and then shut
down from there, which is something that should be changed upstream
in PackageKit. Regardless, it's better than what we have now, and
it's not a big change to set it to shutdown and then apply updates
on the next boot.

Some code modernization and clean-up was also done in shutdowndlg.cpp.

See also https://invent.kde.org/plasma/discover/-/merge_requests/899
which exposes this functionality from Discover.

FEATURE: 452820
FIXED-IN: 6.2.0