applets/systemtray: Use more international-friendly grid layout
The current square grid layout in the hidden items view does not work
well for languages with long strings and especially long words. Let's
fix this by putting the icon next to the text and having only two
columns, leaving more room for text.

This is a good compromise between the current layout and the old list
view, which avoided the problem but was not very space-efficient or
touch-friendly, and left a lot of ugly empty space on the right.