applets/systemtray: stop overriding delegates' tooltip locations
In Plasma 5, this was done to work around some positioning bugs in
PlasmaCore.ToolTipArea. However these have since been fixed, and
overriding the position now causes new bugs, such as the tooltip
appearing under the mouse pointer and appearing and disappearing in a
loop, which is quite hilarious.

Removing this code causes the tooltip to appear below the delegate,
which is fine.

BUG: 470650
BUG: 490555
FIXED-IN: 6.1.5
(cherry picked from commit 7f832e82eb02c5a5a28e728659cca30639bf5106)