dataengines/weather: Fix stalled data on weather updates
Due to the weather's ion and the DataEngine's infrastructure it was
not possible to remove outdated keys from the data, which resulted
in the applet showing stalled info that were no longer relevant, like
"wind gust" or the "feels like" indices.

The root cause is double:

- A dataengine's `setData()` will not remove keys that are no longer
  present in the new data, so we need to force removing them manually
  via `removeData()` or `removeAllData()`

- There are two concurrent dataengines involved:
  - The master `weather` engine which the applet is connected to and
    serves as a general interface
  - An internal `ion` engine loaded by the master which does the
    actual weather job

To fix the issue let's provide a signal for the ions to request
cleaning up the data on the master engine, and call it before setting
the new weather updates.

BUG: 481492
FIXED-IN: 6.2