Make HistoryItemPtr std::unique_ptr, rename HistoryItemConstPtr to HistoryItemSharedPtr
We need HistoryItemPtr to be a unique_ptr because we only make HistoryItem
from HistoryItem::create() most of the time. Adding HistoryItem to a row in
QTest requires raw pointers because QTest will try to append a pointer to a
smart pointer instead of actually appending a smart pointer. shared_ptr
can't release ownership, but std::unique_ptr can be converted to a
shared_ptr and can release ownership.

HistoryItemConstPtr was renamed because it wouldn't be clear that it was a
shared_ptr while HistoryItemPtr is now a unique_ptr.