kcms/users: Add feature to crop avatar images
This is a basic implementation of avatar cropping. On GNOME, the user
is able to crop an image when setting an avatar picture. This tries to
emulate that. thought integrating this into Plasma would be beneficial
as users wouldn't have to make crops in external applications when
setting an avatar.

In the backend, the way this works is that once cropped image is set,
it creates a cropped version of the image file in `/tmp/`. When the
user saves their settings, this image is copied to their account
settings (so, no worry about the file disappearing). Not the best
solution I think, but it works. I'm open for alternatives in the future.