klipper: avoid resizing loops in popup due to screen changes
When the popup opens at a position where two screens with different
resolutions can see the popup, the belonging screen of the popup might
change, so the size is also updated. There is a problem when a new
window size also changes the belonging screen of the popup, which causes
an infinite resizing loop.