applets/activitybar: add activity manager action
The Activity Pager applet has a context menu action that opens the
Activity Switcher, but the Activity Bar applet does not. For consistency
across applets doing similar things, it would be good to have them match.

This change adds such a context menu action to activitybar, placed in
the same position as in activitypager.