Calendar/DayDelegate: Use plain Label for day numbers
It was using a Kirigami.Heading but overriding the pixelSize. This was
not compatible because Heading internally sets the pointSize, leading
to a storm of "Both point size and pixel size set. Using pixel size."
warnings.

Since we're overriding every element, there's no advantage to using a
Heading anyway, and semantically it isn't a Heading. So use a plain
Label.