waylandtasksmodel: don't remove transient window from windows list
Removing window from windows list causes the window to be destroyed. As
alternative, use the SkipTaskbar role to hide the window.

If PlasmaWindow is destroyed, PlasmaWindow::parentWindowChanged is no longer trigged. So it is no longer possible that Window go back do taskbar if it lose their parent.

To avoid this, use the SkipTaskbar to remove window from taskbar.

Also a missing transients.insert(window, leader) was added.

BUG: 484647
BUG: 488596