klipper: guard against null in ClipboardMenu
The ClipboardMenu's focus state depends on the window's visibility, but
this fails when the window doesn't exist yet (such as on plasmashell
start).

This change introduces a guard against this, removing an unnecessary
warning in plasmashell output. It also adds a comment regarding the
purpose of checking for visibility here.