kcmregionandlang.cpp: Clear baseLocaleMap strings from ".UTF-8" and similar text
When constructing the baseLocaleMap, we also have the
.UTF-8 after the glibcLocale in the map.
Later in "fullMatch" iterator this breaks the matching,
since we're comparing for example "en_GB" to "en_GB.UTF-8".

They're both the same thing, but since the strings do not match,
we do not get the exact values and end up doing a lot of extra
work to match the locale to closest possible one with heuristics,
even though we have exact match already.

Removing the .UTF-8 and other similar parts from the prefixedLocales
list fixes the issue and allows full matching.

BUG:461904
FIXED-IN: 6.2.0