lookandfeel: Explicitly set theme to Breeze in defaults
Currently, we only set "library" to "org.kde.breeze", which means
for "theme" the fallback mechanism is used. Depending on how
plasma-workspace was compiled, this is either Breeze or
kwin4_decoration_qml_plastik.

When plasma-workspace was built without Breeze, we end up
with library=org.kde.breeze, theme=kwin4_decoration_qml_plastik,
which is not a valid combination. While this will ultimately result
in Breeze decoration being used as intended, it breaks the
decoration KCM highlight since it tries to select a combination
of library/theme that doesn't exist. KWin might also get confused,
its support information also prints this bogus configuration.

While this doesn't fix the underlying issue of defaults handling
in lnf (see also !3345) and might also indicate a packaging issue
of plasma-workspace, it is imho still advisable to be explicit.

Signed-off-by: Sebastian Stark <sebastian.stark@advantest.com>