Only show logout UI on the screen with the pointer
Otherwise one appears on every screen, and this is awkward-looking. We
have a lot of user complaints about it.

But we don't want to force it to only a particular screen, since if that
screen is turned off but not unplugged, it may still be sending a signal
to the OS and show up as a valid screen to show things on.

To alleviate that concern, always show the UI on the screen that has the
pointer on it.