Fix potential overflow of agenda heading of digital clock
The original approach to determining the heading height is to just use
the height of the MonthView and hope it would be enough. This
assumption is untrue for many font/theme combinations and can cause
spacing issues even if the height happens to be enough.

We fix this by synchronizing the height of both headers to be the
larger one.

BUG: 493344


(cherry picked from commit c083f4674fe09aeaad4fc129604749ee5e710f80)

Co-authored-by: tusooa the west wind <tusooa@kazv.moe>