kcm/cursortheme: Hide themes without their own cursor images
Although those themes (inheriting other valid cursor themes)
can be used as a cursor theme, for the purpose of this KCM,
they are duplicate of the theme they inherit.

The downside of this change is when the cursor theme is set
via other means (e.g. global theme), then it's possible that
the current theme won't show up in this KCM.