kcms/cursortheme: fix the default theme
If the config assumes that the default cursor theme is called "breeze_cursors", then it removes
the theme entry from the config file if you select breeze. However, the default is not necessarily
breeze, it's whatever the distro configures as the "default" theme, so other software might not load
the breeze cursor when the "default" theme is set.
To fix that, this commit changes the default for the config to "default", so that the breeze key is
always written.


(cherry picked from commit f5f27f5fc6b81773391ca5df92f25944be32caaf)

Co-authored-by: Xaver Hugl <xaver.hugl@gmail.com>