NotificationItem: Fix freeze with notifications that have multiple buttons
The Item visibility property that fixes the gap here causes a bindingloop that can freeze
plasmashell since its comparing width and implicitWidth.

The gap is fixed by giving the buttons smallSpacing as a leftMargin, and setting spacing to 0. We also use the leftMargin only for items that are not in first index.