fingerprintmodel.cpp: Fix spamming signals if changing to same value
If any "set" method was called, such as "setDialogState", we
did not check for the previous state and just changed it,
then signaled the change.

This causes the FingerprintDialog StackView to get upset and
show a lot of warnings how it's not done yet.

Only change the state if it's not the same as the previous one to
avoid this issue.