applets/notifications: Port away from Plasma::Dialog
Plasma::Dialog is deprecated, it was a class that tried to do everything
and did everything badly, especially sizing.

The new approach has only a thin base PlasmaWindow that only renders a
frame and the main content with subclasses providing domain specific
integration.

This patch keeps notifications using absolute positioning. Something that might not be wanted long term in public API but is the same as the current state.

There should be no behavioural changes, other than hopefully fixing some sizing issues.