applets/devicenotifier: avoid duplicate entries in device list
When a device is removed and added back immediately, the delay remove
timer is still running, which means in m_devices the old device entry
still exists. If the device is removed later, there is a chance that the
associated ActionsControl is also deleted twice. To make sure that
doesn't happen, always check if a timer is already running when a new
device is added, if so, activate the timer to remove the old device.

SENTRY: PLASMA-WORKSPACE-13P4


(cherry picked from commit 362e95528bd6f841185afaabdaa7d33bd6b8ad83)

Co-authored-by: Fushan Wen <qydwhotmail@gmail.com>