kcm/cursortheme: Hide themes without their own cursor images
Although those themes (inheriting other valid cursor themes)
can be used as a cursor theme, for the purpose of this KCM,
they are duplicates of the theme they inherit.

The current theme is always visible, to prevent the KCM showing
no theme selected on launch. The default theme is also always
visible, so the user can go back to the default at any time.