startkde: Avoid waiting when notifying ksplash is up
During startup we send signals to ksplash to let it knows components are up.
To avoid a previous race ksplash uses the "waitforname" trick so any calls to
notify ksplash before ksplash is up wait around before being dispatched.

This unfortunately means if ksplash is disabled or explicitly closed we still
hit this path and any call to ksplash will be kept alive until timeout.

This systemd service is just used to notify ksplash, we don't need to wait for
a response for the service to be considered done. This can block the login
process, which has the side effect of inhibiting logout too.

BUG: 357170
BUG: 492492