FingerprintDialog: Place ColumnLayout with the stack inside contentItem
This MR has two commits that affect the behavior of the dialog.

* FingerprintDialog: Place ColumnLayout with the stack inside contentItem

Previously we set the implicitSize directly to the dialog. This causes weird issues since
the stackView can be larger than the dialog. Setting the same values as the minimum values of
the stackView fixes the issue, since the size of the dialog is calculated through the stackView
values (since it's the dialogs content). Also fix some widths being applied into implicitWidth properties.

Requires https://invent.kde.org/frameworks/kirigami/-/merge_requests/1637

BUG:493977
BUG:493978
BUG:493976


* fingerprintmodel.cpp: Fix spamming signals if changing to same value

If any "set" method was called, such as "setDialogState", we
did not check for the previous state and just changed it,
then signaled the change.

This causes the FingerprintDialog StackView to get upset and
show a lot of warnings how it's not done yet.

Only change the state if it's not the same as the previous one to
avoid this issue.