applets/devicenotifier: fix crash when parent device is not of StorageDrive type
If a storage volume's parent is not a storage device, the parent will
not be added to m_parentDevices. When the storage volume is unmounted,
we also need to check if the parent exists in m_parentDevices before
using it.

SENTRY: PLASMA-WORKSPACE-146Y
FIXED-IN: 6.2.1


(cherry picked from commit 009dad822f69fbf2bce7e58c0cf4cd67feeddfc9)

Co-authored-by: Fushan Wen <qydwhotmail@gmail.com>