startkde: Specify --type=method_call to correctly notify ksplash
In 3e295c33d0 (startkde: Avoid waiting when notifying ksplash is up,
2024-10-11) the invocation of the dbus interface was changed from qdbus
to dbus-send. The former defaults to calling a method, whilst the latter
defaults to sending a signal, causing the setStage method to never be
called.

Instead of using the default behaviour, specify explicitly that we want
to call a method using --type=method_call.

CCBUG: 494840