klipper: Avoid creating lots of native windows in settings dialog
Klipper's action dialog tries to resize itself, for that it needs a
QWindow. It was using an extreme call that also makes it a native window
ahead of QDialog doing it.

The dialog has a widget parent which isn't separated yet due to it not
being shown.

The current code hits a path where QWidget tries making all child
widgets of the parent widget also native windows and we get in a mess.

BUG: 494169


(cherry picked from commit 44358a8e724aeba3a4520be9ae63d6c4fec4c529)

Co-authored-by: David Edmundson <kde@davidedmundson.co.uk>