klipper: clear search field after a resonable amount of time
When a user opens the klipper popup menu, they might want to get the
latest unfiltered result instead of the last search result because
the clipboard stores many temporary entries, so any filtered result is
only useful for a certain period.

This makes the search field cleared after the menu is closed. Before the
search field is cleared, a reasonable amount of time is given to keep
the search result in case the menu is closed accidentally.