weather/dwd: Add support for night-time icons
The DWD Ion currently shows only day-time icons which feels a bit off
when checking the weather after dusk or at night. Implement the missing
night-time icons and use sunrise and sunset times from DWD's forecast
endpoint to accurately determine when to switch. This will only affect
the display of the current conditions, the forecast icons will continue
to be displayed as day-time only.

Since we need to have both observation data (for the current observation
time) and forecast data (for sunrise and sunset times), only determine
the icon at a very late stage in updateWeather(). Additionally, make
sure to safely fall back to day-time only icons if we could not fetch
sunrise or sunset times.