applets/devicenotifier: ensure udi is not dangling to fix crash
udi in DeviceControl::deviceDelayRemove is passed as const reference, so
the inner refcount will not be increased, which causes a problem that if
udi is from m_removeTimers, it will become dangling. This ensures the
timer data is removed at the end of DeviceControl::deviceDelayRemove to
fix the crash.

As a drive-by, in the timer, checking if udi exists in m_devices is unnecessary
because DeviceControl::deviceDelayRemove also checks that.

SENTRY: PLASMA-WORKSPACE-15SV
FIXED-IN: 6.2.2


(cherry picked from commit cf9fb7cbc66fd83c2ec4c5346faf459a5d0f5a72)

Co-authored-by: Fushan Wen <qydwhotmail@gmail.com>