Fix default keybindings clash
### Reason for the change

<!--
If this merge request change is resolving a problem, describe the problem and explain why the approach taken here is the correct way to resolve it.

If the merge request adds a new feature or UI change, explain the benefits of the change.
-->

Currently, the default shortcuts for "Activate Task Manager Entry 10" and "Zoom to Actual Size" clash with each other. Also "Manually Invoke Action on Current Clipboard" and "Record Window" clash.

Remove the default shortcut of "Activate Task Manager Entry 10" & "Manually Invoke Action on Current Clipboard" to fix this bug.

### Test plan

<!--
Extend any related autotests with new test cases to validate your changes. If none exist, consider adding one if you have the technical skills to do so.

Otherwise, at least describe how you tested the change, and how someone reviewing this merge request can test it themselves. No need to mention obvious things like "Apply the changes".
-->

### Screenshots or screen recordings

<!--
If this merge request introduces a visual change, please add before-and-after screenshots in the following format:

| Before | After |
| ------ | ----- |
| [drag "before" screenshot here] | [drag "after" screenshot here] |
-->

### Bugs fixed

<!--
If the changes in this merge request fix any Bugzilla tickets, add the following keyword for each one:

BUG: [number of the bug report]
-->

BUG: 491700